refactor:💡 ScheduleResponse dto 패키지 하위로 이동
