+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/GetTodayScheduleByParentsResponse.java
@@ -1,6 +1,7 @@
 package gwasuwonshot.tutice.schedule.dto.response;
 
 import gwasuwonshot.tutice.common.dto.GetTodayDateResponse;
+import gwasuwonshot.tutice.schedule.dto.ScheduleResponse;
 import gwasuwonshot.tutice.schedule.entity.Schedule;
 import lombok.*;
