+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/GetTodayScheduleByTeacherResponse.java
@@ -1,5 +1,6 @@
 package gwasuwonshot.tutice.schedule.dto.response;
 
+import gwasuwonshot.tutice.schedule.dto.ScheduleResponse;
 import gwasuwonshot.tutice.schedule.entity.Schedule;
 import lombok.*;
