+++ src/main/java/gwasuwonshot/tutice/schedule/dto/response/LatestScheduleByTeacher.java
@@ -1,6 +1,7 @@
 package gwasuwonshot.tutice.schedule.dto.response;
 
 import gwasuwonshot.tutice.lesson.entity.Lesson;
+import gwasuwonshot.tutice.schedule.dto.ScheduleResponse;
 import gwasuwonshot.tutice.schedule.entity.Schedule;
 import lombok.*;
